requirements.txt
Django==4.2.7
djangorestframework==3.15.0
# mysqlclient==2.2.4
gunicorn==21.2.0
whitenoise==6.5.0
Faker==30.3.0
python-decouple==3.8
psycopg2-binary>=2.9.7,<2.10.0
dj-database-url>=2.1,<2.2
django-debug-toolbar>=3.8.1,<3.9.0
django-redis>=5.3.0,<=5.4.0
django-cors-headers>=3.7.0,<4.3.1
djangorestframework-jwt
PyJWT
django-filter
pillow
pytz
